requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Vector Database
qdrant-client==1.7.0

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Environment and Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# OpenAI
openai==1.12.0

# Note: Install dependencies using:
# pip install -r requirements.txt
#
# Run the server locally using:
# uvicorn main:app --reload 
